Add display rotation support
Add support for rotating the display the 4 standard orientations.  It
uses the screen16 array (pointer) to compute the right pixel to update
depending on current _width and _height.

Also fixed some issues assocated with updateOnce.  Needed to clear flag
so the BeginTransaction code would be called again.  Likewise needed to
disable the DMA write after we call stopRefresh